RL training runs, now need to debug the values. also updated the metrics

+++ source/robot_rl/robot_rl/tasks/manager_based/robot_rl/mdp/commands/traj_tracking/library_manager.py
@@ -152,7 +152,7 @@ class LibraryManager:
 
         # Initialize output tensor
         N = t.shape[0]
-        frame_indices = torch.zeros(N, device=self.device)
+        frame_indices = torch.zeros(N, dtype=torch.long, device=self.device)
 
         # Get the unique managers (avoid repeats)
         unique_indicies = torch.unique(indices)
